+++ mcl/Measurement.h
@@ -6,6 +6,6 @@ namespace mcl {
         int measured_landmark_id;
         int gt_landmark_id;
         float u, v;
-        Eigen::Matrix<float, mcl::APPEARANCE_SIZE, 1> appearance;
+        Eigen::Matrix<double, mcl::APPEARANCE_SIZE, 1> appearance;
     };
 }
